Migrate more tests and add agora-test-context
