+++ Cargo.toml
@@ -9,7 +9,7 @@ repository = "https://github.com/agora-org/agora"
 homepage = "https://github.com/agora-org/agora"
 
 [workspace]
-members = [".", "agora-lnd-client", "bin/prerelease", "bin/publish", "lnd-test-context"]
+members = [".", "agora-lnd-client", "agora-test-context", "bin/prerelease", "bin/publish", "lnd-test-context"]
 
 [dependencies]
 color-backtrace = "0.5.1"
@@ -83,6 +83,9 @@ tiny-skia = "0.5.1"
 unindent = "0.1.7"
 usvg = "0.15.0"
 
+[dev-dependencies.agora-test-context]
+path = "agora-test-context"
+
 [dev-dependencies.lnd-test-context]
 path = "lnd-test-context"
 
@@ -94,10 +97,6 @@ features = ["x509-parser"]
 version = "0.11.3"
 features = ["blocking", "stream", "rustls-tls"]
 default-features = false
-
-[[test]]
-name = "integration"
-path = "tests/integration.rs"
 
 [features]
 slow-tests = []
