+++ agora-test-context/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "agora-test-context"
+version = "0.0.0"
+edition = "2021"
+publish = false
+
+[dependencies]
+tempfile = "3.2.0"
+executable-path = "1.0.0"
+
+[dependencies.reqwest]
+version = "0.11.3"
+features = ["blocking", "stream", "rustls-tls"]
+default-features = false
