+++ src/tests.rs
@@ -1,6 +1,5 @@
 use crate::{
   common::*,
-  server::TestContext,
   test_utils::{
     assert_contains, assert_not_contains, https_client, set_up_test_certificate, test,
     test_with_arguments, test_with_environment,
@@ -8,7 +7,7 @@ use crate::{
 };
 use guard::guard_unwrap;
 use pretty_assertions::assert_eq;
-use reqwest::{redirect::Policy, Client, Url};
+use reqwest::Url;
 use scraper::{ElementRef, Html, Selector};
 use std::{net::TcpListener, path::MAIN_SEPARATOR};
 use tokio::net::TcpStream;
@@ -35,6 +34,26 @@ async fn html(url: &Url) -> Html {
 fn css_select<'a>(html: &'a Html, selector: &'a str) -> Vec<ElementRef<'a>> {
   let selector = Selector::parse(selector).unwrap();
   html.select(&selector).collect::<Vec<_>>()
+}
+
+fn symlink(contents: impl AsRef<Path>, link: impl AsRef<Path>) {
+  #[cfg(unix)]
+  std::os::unix::fs::symlink(contents, link).unwrap();
+  #[cfg(windows)]
+  {
+    let target = link.as_ref().parent().unwrap().join(&contents);
+    if target.is_dir() {
+      std::os::windows::fs::symlink_dir(contents, link).unwrap();
+    } else if target.is_file() {
+      std::os::windows::fs::symlink_file(contents, link).unwrap();
+    } else {
+      panic!(
+        "unsupported file type for paths: contents: `{}`, link: `{}`",
+        contents.as_ref().display(),
+        link.as_ref().display(),
+      );
+    }
+  }
 }
 
 #[test]
@@ -66,28 +85,6 @@ fn configure_port() {
         .status(),
       200
     )
-  });
-}
-
-#[test]
-fn unknown_route_status_code_is_404() {
-  test(|context| async move {
-    assert_eq!(
-      reqwest::get(context.base_url().join("huhu").unwrap())
-        .await
-        .unwrap()
-        .status(),
-      404
-    )
-  });
-}
-
-#[test]
-fn index_route_contains_title() {
-  test(|context| async move {
-    let haystack = text(context.base_url()).await;
-    let needle = "<title>agora</title>";
-    assert_contains(&haystack, needle);
   });
 }
 
@@ -129,87 +126,6 @@ fn errors_in_request_handling_cause_500_status_codes() {
     &stderr,
     "IO error accessing filesystem at `www/foo`: Permission denied (os error 13)",
   );
-}
-
-#[test]
-fn listing_contains_file() {
-  test(|context| async move {
-    context.write("some-test-file.txt", "");
-    let haystack = html(context.base_url()).await.root_element().html();
-    let needle = "some-test-file.txt";
-    assert_contains(&haystack, needle);
-  });
-}
-
-#[test]
-fn listing_contains_multiple_files() {
-  test(|context| async move {
-    context.write("a.txt", "");
-    context.write("b.txt", "");
-    let haystack = html(context.base_url()).await.root_element().html();
-    assert_contains(&haystack, "a.txt");
-    assert_contains(&haystack, "b.txt");
-  });
-}
-
-#[test]
-fn listing_is_sorted_alphabetically() {
-  test(|context| async move {
-    context.write("b", "");
-    context.write("c", "");
-    context.write("a", "");
-    let html = html(context.base_url()).await;
-    let haystack: Vec<&str> = css_select(&html, ".listing a:not([download])")
-      .into_iter()
-      .map(|x| x.text())
-      .flatten()
-      .collect();
-    assert_eq!(haystack, vec!["a", "b", "c"]);
-  });
-}
-
-#[test]
-fn listed_files_can_be_played_in_browser() {
-  test(|context| async move {
-    context.write("some-test-file.txt", "contents");
-    let html = html(context.files_url()).await;
-    guard_unwrap!(let &[a] = css_select(&html, ".listing a:not([download])").as_slice());
-    assert_eq!(a.inner_html(), "some-test-file.txt");
-    let file_url = a.value().attr("href").unwrap();
-    let file_url = context.files_url().join(file_url).unwrap();
-    let file_contents = text(&file_url).await;
-    assert_eq!(file_contents, "contents");
-  });
-}
-
-#[test]
-fn listed_files_have_download_links() {
-  test(|context| async move {
-    context.write("some-test-file.txt", "contents");
-    let html = html(context.files_url()).await;
-    guard_unwrap!(let &[a] = css_select(&html, "a[download]").as_slice());
-    assert_contains(&a.inner_html(), "download");
-    let file_url = a.value().attr("href").unwrap();
-    let file_url = context.files_url().join(file_url).unwrap();
-    let file_contents = text(&file_url).await;
-    assert_eq!(file_contents, "contents");
-  });
-}
-
-#[test]
-fn listed_files_have_percent_encoded_hrefs() {
-  test(|context| async move {
-    context.write("filename with special chäracters", "");
-    let html = html(context.base_url()).await;
-    let links = css_select(&html, ".listing a");
-    assert_eq!(links.len(), 2);
-    for link in links {
-      assert_eq!(
-        link.value().attr("href").unwrap(),
-        "filename%20with%20special%20ch%C3%A4racters"
-      );
-    }
-  });
 }
 
 #[test]
@@ -280,16 +196,6 @@ fn return_404_for_missing_files() {
 }
 
 #[test]
-fn serves_error_pages() {
-  test(|context| async move {
-    let response = reqwest::get(context.files_url().join("foo.txt").unwrap())
-      .await
-      .unwrap();
-    assert_contains(&response.text().await.unwrap(), "404 Not Found");
-  });
-}
-
-#[test]
 fn configure_source_directory() {
   let mut environment = Environment::test();
   environment.arguments = vec![
@@ -312,106 +218,6 @@ fn configure_source_directory() {
 }
 
 #[test]
-#[cfg(unix)]
-fn downloaded_files_are_streamed() {
-  use futures::StreamExt;
-  use tokio::{fs::OpenOptions, sync::oneshot};
-
-  test(|context| async move {
-    let fifo_path = context.files_directory().join("fifo");
-
-    nix::unistd::mkfifo(&fifo_path, nix::sys::stat::Mode::S_IRWXU).unwrap();
-
-    let (sender, receiver) = oneshot::channel();
-
-    let writer = tokio::spawn(async move {
-      let mut fifo = OpenOptions::new()
-        .write(true)
-        .open(&fifo_path)
-        .await
-        .unwrap();
-      fifo.write_all(b"hello").await.unwrap();
-      receiver.await.unwrap();
-    });
-
-    let mut stream = get(&context.files_url().join("fifo").unwrap())
-      .await
-      .bytes_stream();
-
-    assert_eq!(stream.next().await.unwrap().unwrap(), "hello");
-
-    sender.send(()).unwrap();
-
-    writer.await.unwrap();
-  });
-}
-
-#[test]
-fn downloaded_files_have_correct_content_type() {
-  test(|context| async move {
-    context.write("foo.mp4", "hello");
-
-    let response = get(&context.files_url().join("foo.mp4").unwrap()).await;
-
-    assert_eq!(
-      response.headers().get(header::CONTENT_TYPE).unwrap(),
-      "video/mp4"
-    );
-  });
-}
-
-#[test]
-fn unknown_files_have_no_content_type() {
-  test(|context| async move {
-    context.write("foo", "hello");
-
-    let response = get(&context.files_url().join("foo").unwrap()).await;
-
-    assert_eq!(response.headers().get(header::CONTENT_TYPE), None);
-  });
-}
-
-#[test]
-fn filenames_with_spaces() {
-  test(|context| async move {
-    context.write("foo bar", "hello");
-
-    let response = text(&context.files_url().join("foo%20bar").unwrap()).await;
-
-    assert_eq!(response, "hello");
-  });
-}
-
-#[test]
-fn subdirectories_appear_in_listings() {
-  test(|context| async move {
-    context.write("foo/bar.txt", "hello");
-    let root_listing = html(context.files_url()).await;
-    guard_unwrap!(let &[a] = css_select(&root_listing, ".listing a").as_slice());
-    assert_eq!(a.inner_html(), "foo/");
-    let subdir_url = context
-      .files_url()
-      .join(a.value().attr("href").unwrap())
-      .unwrap();
-    let subdir_listing = html(&subdir_url).await;
-    guard_unwrap!(let &[a] = css_select(&subdir_listing, ".listing a:not([download])").as_slice());
-    assert_eq!(a.inner_html(), "bar.txt");
-    let file_url = subdir_url.join(a.value().attr("href").unwrap()).unwrap();
-    assert_eq!(text(&file_url).await, "hello");
-  });
-}
-
-#[test]
-fn redirects_correctly_for_two_layers_of_subdirectories() {
-  test(|context| async move {
-    context.write("foo/bar/baz.txt", "");
-    let listing = html(&context.files_url().join("foo/bar").unwrap()).await;
-    guard_unwrap!(let &[a] = css_select(&listing, ".listing a:not([download])").as_slice());
-    assert_eq!(a.inner_html(), "baz.txt")
-  });
-}
-
-#[test]
 fn file_errors_are_associated_with_file_path() {
   let stderr = test(|context| async move {
     fs::create_dir(context.files_directory().join("foo")).unwrap();
@@ -430,79 +236,6 @@ fn file_errors_are_associated_with_file_path() {
       MAIN_SEPARATOR, MAIN_SEPARATOR,
     ),
   );
-}
-
-#[test]
-fn requesting_files_with_trailing_slash_redirects() {
-  test(|context| async move {
-    context.write("foo", "");
-    let response = reqwest::get(context.files_url().join("foo/").unwrap())
-      .await
-      .unwrap();
-    assert!(
-      response.url().as_str().ends_with("/files/foo"),
-      "{} didn't end with /files/foo",
-      response.url()
-    );
-  });
-}
-
-#[test]
-fn listings_are_not_cached() {
-  test(|context| async move {
-    let response = reqwest::get(context.files_url().clone()).await.unwrap();
-    assert_eq!(
-      response.headers().get(header::CACHE_CONTROL).unwrap(),
-      "no-store, max-age=0",
-    );
-  });
-}
-
-#[test]
-fn files_are_not_cached() {
-  test(|context| async move {
-    context.write("foo", "bar");
-    let response = reqwest::get(context.files_url().join("foo").unwrap())
-      .await
-      .unwrap();
-    assert_eq!(
-      response.headers().get(header::CACHE_CONTROL).unwrap(),
-      "no-store, max-age=0",
-    );
-    assert_eq!(response.text().await.unwrap(), "bar");
-  });
-}
-
-fn symlink(contents: impl AsRef<Path>, link: impl AsRef<Path>) {
-  #[cfg(unix)]
-  std::os::unix::fs::symlink(contents, link).unwrap();
-  #[cfg(windows)]
-  {
-    let target = link.as_ref().parent().unwrap().join(&contents);
-    if target.is_dir() {
-      std::os::windows::fs::symlink_dir(contents, link).unwrap();
-    } else if target.is_file() {
-      std::os::windows::fs::symlink_file(contents, link).unwrap();
-    } else {
-      panic!(
-        "unsupported file type for paths: contents: `{}`, link: `{}`",
-        contents.as_ref().display(),
-        link.as_ref().display(),
-      );
-    }
-  }
-}
-
-#[test]
-fn allow_file_downloads_via_local_symlinks() {
-  test(|context| async move {
-    context.write("file", "contents");
-    symlink("file", context.files_directory().join("link"));
-    let response = reqwest::get(context.files_url().join("link").unwrap())
-      .await
-      .unwrap();
-    assert_eq!(response.status(), StatusCode::OK);
-  });
 }
 
 #[test]
@@ -546,18 +279,6 @@ fn disallow_file_downloads_via_absolute_escaping_symlinks() {
 }
 
 #[test]
-fn allow_file_downloads_via_local_intermediate_symlinks() {
-  test(|context| async move {
-    context.write("dir/file", "contents");
-    symlink("dir", context.files_directory().join("link"));
-    let response = reqwest::get(context.files_url().join("link/file").unwrap())
-      .await
-      .unwrap();
-    assert_eq!(response.status(), StatusCode::OK);
-  });
-}
-
-#[test]
 fn disallow_file_downloads_via_escaping_intermediate_symlinks() {
   let stderr = test(|context| async move {
     context.write("../dir/file", "contents");
@@ -574,19 +295,6 @@ fn disallow_file_downloads_via_escaping_intermediate_symlinks() {
       MAIN_SEPARATOR
     ),
   );
-}
-
-#[test]
-fn allow_listing_directories_via_local_symlinks() {
-  test(|context| async move {
-    let dir = context.files_directory().join("dir");
-    fs::create_dir(&dir).unwrap();
-    symlink("dir", context.files_directory().join("link"));
-    let response = reqwest::get(context.files_url().join("link").unwrap())
-      .await
-      .unwrap();
-    assert_eq!(response.status(), StatusCode::OK);
-  });
 }
 
 #[test]
@@ -610,20 +318,6 @@ fn disallow_listing_directories_via_escaping_symlinks() {
 }
 
 #[test]
-fn allow_listing_directories_via_intermediate_local_symlinks() {
-  test(|context| async move {
-    let dir = context.files_directory().join("dir");
-    fs::create_dir(&dir).unwrap();
-    symlink("dir", context.files_directory().join("link"));
-    fs::create_dir(dir.join("subdir")).unwrap();
-    let response = reqwest::get(context.files_url().join("link/subdir").unwrap())
-      .await
-      .unwrap();
-    assert_eq!(response.status(), StatusCode::OK);
-  });
-}
-
-#[test]
 fn disallow_listing_directories_via_intermediate_escaping_symlinks() {
   let stderr = test(|context| async move {
     let dir = context.files_directory().join("../dir");
@@ -642,107 +336,6 @@ fn disallow_listing_directories_via_intermediate_escaping_symlinks() {
       MAIN_SEPARATOR
     ),
   );
-}
-
-#[test]
-fn show_local_symlinks_in_listings() {
-  test(|context| async move {
-    context.write("file", "");
-    symlink("file", context.files_directory().join("link"));
-    let html = html(context.files_url()).await;
-    guard_unwrap!(let &[a, b] = css_select(&html, ".listing a:not([download])").as_slice());
-    assert_eq!(a.inner_html(), "file");
-    assert_eq!(b.inner_html(), "link");
-  });
-}
-
-#[test]
-fn remove_escaping_symlinks_from_listings() {
-  test(|context| async move {
-    context.write("../escaping", "");
-    context.write("local", "");
-    symlink("../escaping", context.files_directory().join("link"));
-    let html = html(context.files_url()).await;
-    guard_unwrap!(let &[a] = css_select(&html, ".listing a:not([download])").as_slice());
-    assert_eq!(a.inner_html(), "local");
-  });
-}
-
-#[test]
-fn serves_static_assets() {
-  test(|context| async move {
-    let response = text(&context.base_url().join("static/index.css").unwrap()).await;
-    let expected = fs::read_to_string("static/index.css").unwrap();
-    assert_eq!(response, expected);
-  });
-}
-
-#[test]
-fn sets_mime_types_for_static_assets() {
-  test(|context| async move {
-    let response = get(&context.base_url().join("static/index.css").unwrap()).await;
-    assert_eq!(
-      response.headers().get(header::CONTENT_TYPE).unwrap(),
-      "text/css"
-    );
-  });
-}
-
-#[test]
-fn missing_asset_not_found() {
-  test(|context| async move {
-    let response = reqwest::get(context.base_url().join("static/does-not-exist").unwrap())
-      .await
-      .unwrap();
-    assert_eq!(response.status(), StatusCode::NOT_FOUND);
-  });
-}
-
-#[test]
-fn listing_does_not_contain_hidden_file() {
-  test(|context| async move {
-    context.write(".some-test-file.txt", "");
-    let haystack = html(context.base_url()).await.root_element().html();
-    let needle = ".some-test-file.txt";
-    assert_not_contains(&haystack, needle);
-  });
-}
-
-#[test]
-fn return_404_for_hidden_files() {
-  test(|context| async move {
-    context.write(".foo.txt", "");
-    assert_eq!(
-      reqwest::get(context.files_url().join(".foo.txt").unwrap())
-        .await
-        .unwrap()
-        .status(),
-      StatusCode::NOT_FOUND
-    )
-  });
-}
-
-#[test]
-fn return_404_for_hidden_directories() {
-  test(|context| async move {
-    let dir = context.files_directory().join(".dir");
-    fs::create_dir(&dir).unwrap();
-    let response = reqwest::get(context.files_url().join(".dir").unwrap())
-      .await
-      .unwrap();
-    assert_eq!(response.status(), StatusCode::NOT_FOUND);
-  });
-}
-
-#[test]
-fn return_404_for_files_in_hidden_directories() {
-  test(|context| async move {
-    context.write(".dir/foo.txt", "hello");
-    let response = reqwest::get(context.files_url().join(".dir/foo.txt").unwrap())
-      .await
-      .unwrap();
-    assert_eq!(response.status(), StatusCode::NOT_FOUND);
-  });
 }
 
 #[test]
@@ -958,28 +551,6 @@ fn redirects_requests_from_port_80_to_443() {
       assert_eq!(body, "encrypted content");
     },
   );
-}
-
-#[test]
-fn favicon_is_served_at_favicon_ico() {
-  test(|context| async move {
-    let response = get(&context.base_url().join("favicon.ico").unwrap()).await;
-    assert_eq!(
-      response.headers().get(header::CONTENT_TYPE).unwrap(),
-      "image/x-icon"
-    );
-  });
-}
-
-#[test]
-fn apple_touch_icon_is_served_under_root() {
-  test(|context| async move {
-    let response = get(&context.base_url().join("apple-touch-icon.png").unwrap()).await;
-    assert_eq!(
-      response.headers().get(header::CONTENT_TYPE).unwrap(),
-      "image/png"
-    );
-  });
 }
 
 #[test]
